fix(config): eslint/prettier/ts configs
Add build output to eslint / prettier configs, and remove prettier
arrowParens option (falling back to the prettier 2 deafult). the
typescript config file was modified to add comments for each option so
people less familiar with the config can understand it better. i left in
the duplicate / unecessary options inside the config, not opting to
remove them